requirements.txt
# This file is used to install the required packages for the project.
seaborn
matplotlib
pandas
numpy
pytest
python-dotenv
ipykernel
groq
requests
setuptools
jupyter
matplotlib>=3.8.0
pytest-cov
pytest-mock

